# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GymPal Backend API",
    "description": "Complete API for the GymPal fitness application with social features, personal tracking, workouts, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "GymPal Team",
      "email": "support@gympal.app",
      "url": "https://gympal.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.gympal.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get user settings",
        "description": "Retrieve all settings for authenticated user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update user settings",
        "description": "Update general settings for authenticated user",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/notifications": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get notification settings",
        "description": "Retrieve notification preferences",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Notification settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update notification settings",
        "description": "Update notification preferences",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification settings updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/privacy": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get privacy settings",
        "description": "Retrieve privacy settings",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Privacy settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update privacy settings",
        "description": "Update privacy settings",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Privacy settings updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/preferences": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get user preferences",
        "description": "Retrieve user preferences",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User preferences retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update user preferences",
        "description": "Update user preferences",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User preferences updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/fitness": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get fitness settings",
        "description": "Retrieve fitness-related settings",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Fitness settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update fitness settings",
        "description": "Update fitness-related settings",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitnessSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fitness settings updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/social": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get social settings",
        "description": "Retrieve social-related settings",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Social settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update social settings",
        "description": "Update social-related settings",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Social settings updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/export-data": {
      "post": {
        "tags": ["Settings"],
        "summary": "Export user data",
        "description": "Export all user data",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "format": {
                    "type": "string",
                    "enum": ["json", "csv"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data exported successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/import-data": {
      "post": {
        "tags": ["Settings"],
        "summary": "Import user data",
        "description": "Import user data",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data imported successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/account": {
      "delete": {
        "tags": ["Settings"],
        "summary": "Delete user account",
        "description": "Permanently delete user account",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "confirmation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/settings/activity-log": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get activity log",
        "description": "Retrieve user activity log",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity log retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SettingsResponse": {
        "type": "object",
        "properties": {
          "general": {
            "$ref": "#/components/schemas/GeneralSettings"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationSettings"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacySettings"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "fitness": {
            "$ref": "#/components/schemas/FitnessSettings"
          },
          "social": {
            "$ref": "#/components/schemas/SocialSettings"
          }
        }
      },
      "UpdateSettingsRequest": {
        "type": "object",
        "properties": {
          "general": {
            "$ref": "#/components/schemas/GeneralSettings"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationSettings"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacySettings"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "fitness": {
            "$ref": "#/components/schemas/FitnessSettings"
          },
          "social": {
            "$ref": "#/components/schemas/SocialSettings"
          }
        }
      },
      "GeneralSettings": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "example": "en"
          },
          "timezone": {
            "type": "string",
            "example": "UTC"
          },
          "units": {
            "type": "string",
            "enum": ["metric", "imperial"],
            "example": "metric"
          },
          "theme": {
            "type": "string",
            "enum": ["light", "dark", "auto"],
            "example": "dark"
          }
        }
      },
      "NotificationSettings": {
        "type": "object",
        "properties": {
          "email": {
            "type": "boolean",
            "example": true
          },
          "push": {
            "type": "boolean",
            "example": true
          },
          "workoutReminders": {
            "type": "boolean",
            "example": true
          },
          "achievements": {
            "type": "boolean",
            "example": true
          },
          "social": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "PrivacySettings": {
        "type": "object",
        "properties": {
          "profileVisibility": {
            "type": "string",
            "enum": ["public", "friends", "private"],
            "example": "public"
          },
          "workoutVisibility": {
            "type": "string",
            "enum": ["public", "friends", "private"],
            "example": "friends"
          },
          "showEmail": {
            "type": "boolean",
            "example": false
          },
          "showStats": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UserPreferences": {
        "type": "object",
        "properties": {
          "workoutDays": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
            },
            "example": ["monday", "wednesday", "friday"]
          },
          "preferredWorkoutTime": {
            "type": "string",
            "example": "07:00"
          },
          "fitnessGoals": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Build muscle", "Increase strength"]
          }
        }
      },
      "FitnessSettings": {
        "type": "object",
        "properties": {
          "experienceLevel": {
            "type": "string",
            "enum": ["beginner", "intermediate", "advanced", "expert"],
            "example": "intermediate"
          },
          "primaryGoal": {
            "type": "string",
            "example": "Build muscle"
          },
          "workoutFrequency": {
            "type": "integer",
            "example": 4
          },
          "preferredDuration": {
            "type": "integer",
            "example": 60
          }
        }
      },
      "SocialSettings": {
        "type": "object",
        "properties": {
          "allowFriendRequests": {
            "type": "boolean",
            "example": true
          },
          "allowMessages": {
            "type": "boolean",
            "example": true
          },
          "showActivity": {
            "type": "boolean",
            "example": true
          },
          "allowTagging": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}